requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.6
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.4.1
googlesearch-python==1.2.5
groq==0.11.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
multidict==6.0.5
numpy==2.0.2
openai==1.44.0
PyAudio==0.2.14
pydantic==2.9.0
pydantic_core==2.23.2
pypiwin32==223
python-engineio==4.9.1
python-socketio==5.11.4
pyttsx3==2.91
pywin32==306
requests==2.32.3
simple-websocket==1.0.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.10.4
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.10.0
zipp==3.20.1
